requirements.txt
pandas
selenium==4.4.3
chromedriver-py
python-pptx
